更新

+++ create_image_features.py
@@ -2,6 +2,7 @@ import json
 import logging
 import os
 import pandas as pd
+import torch
 from tqdm import tqdm
 import cv2
 import detectron2
@@ -59,7 +60,7 @@ def get_image_info_list():
 
 
 def get_image_features(image_dir):
-    ret = torch.empty(0).cuda()
+    ret = torch.empty(0, dtype=torch.long).cuda()
 
     files = os.listdir(image_dir)
     for file in files:
@@ -67,7 +68,7 @@ def get_image_features(image_dir):
         outputs = predictor(im)
 
         pred_classes = outputs["instances"].pred_classes
-        pred_classes = pred_classes.flatten()
+        pred_classes = pred_classes.long().flatten()
 
         pred_boxes = outputs["instances"].pred_boxes
         box_areas = pred_boxes.area().long().flatten()
@@ -86,7 +87,7 @@ def create_image_features():
     os.makedirs(FEATURES_DIR, exist_ok=True)
 
     for i, image_info in enumerate(tqdm(image_info_list)):
-        image_features = torch.zeros(0).cuda()
+        image_features = torch.zeros(0, dtype=torch.long).cuda()
         if image_info.image_dir != "":
             try:
                 image_features = get_image_features(image_info.image_dir)
